+++ NEWS.md
@@ -92,6 +92,7 @@ the new router layout. Items must be wrapped in a `shiny::tagList()`.
 - `f7TabLink()`: `href` can now be different of `#` if passed in `...` (Default is still `#` when no custom `href` is detected). This allows to use tab links as routable links but with a better rendering in the toolbar/tabbar than classic `f7Link()`.
 - `updateF7App` can now also handle changes in app theme (ios or md), dark mode, and color.
 - `f7Fabs()` has a new argument `global` that can be used to make FABs persistent across different tabs in `f7TabLayout()`. 
+- `f7ExpandableCard()` has a new argument `buttonColor` that can be used to control the color of the close button.
 - Fix various issues in documentation.
 - Include new vignettes.